requirements.txt
spacy==3.7.2
pdfplumber==0.10.3
python-docx==1.1.2
transformers==4.51.1
torch==2.6.0